Capitalize module names everywhere, not just when determining their filename.
This fixes the following bug:

When converting a TypeScript file that contains:
 - a TypeScript module of the same name as the file, but with different casing; and
 - at least one top-level declaration in addition to the module;

the Converter would store this as two separate "modules" in the Converter.modules variable:
 - one with the same casing as the TypeScript module, and containing the contents of the TypeScript module; and
 - one with the same casing as the file, and containing the additional top-level declarations.

The two modules would go on to be written to the same filename, resulting in the second module overwriting the first.

An example of a TypeScript file that causes this bug is: https://github.com/borisyankov/DefinitelyTyped/blob/b3aeee64552e324581fdbe0926f4a96e981e451b/ractive/ractive.d.ts (due to the top-level declaration on line 476).